Backend - Add Comments Controller and paths in tsconfig

+++ backend/src/Controllers/PublicationsControllers/PublicationsController.ts
@@ -1,8 +1,8 @@
 // eslint-disable-next-line no-unused-vars
 import { Request, Response } from 'express'
 
-import knex from '../database/connection'
-import nowDateUTC from '../utils/NowDateUTC'
+import knex from '../../database/connection'
+import nowDateUTC from '../../utils/NowDateUTC'
 
 class PublicationsController {
   async index (req: Request, res: Response) {
@@ -12,10 +12,12 @@ class PublicationsController {
     const publications = await knex('publications')
       .leftJoin('public_likes', 'public_likes.publication_id', '=', 'publications.id')
       .where('publications.user_id', Number(id))
-      .groupBy('publications.id')
       .select('publications.*')
+      .groupBy('publications.id')
       .orderBy('publications.date', 'desc')
-      .count('public_likes.publication_id', { as: 'likes' })
+      .count({
+        likes: 'public_likes.publication_id'
+      })
       .limit(20)
       .offset((Number(page) - 1) * 20)
 
@@ -49,10 +51,19 @@ class PublicationsController {
       user_id: userId
     }
 
-    await knex('publications')
-      .where(data).first().delete()
+    const userPublications = await knex('publications_comments')
+      .where(data).first()
 
-    return res.send('')
+    if (!userPublications) return res.status(401).json({ error: 'Only the user who created the publications can delete them' })
+
+    try {
+      await knex('publications')
+        .where(data).first().delete()
+
+      return res.send('')
+    } catch (e) {
+      return res.status(500).json({ error: 'Error, Try again' })
+    }
   }
 }
